+++ src/types/config/index.ts
@@ -1,2 +1,3 @@
 export * from './AppConfig'
 export * from './ImapAccountConfig'
+export * from './SettingsRow'